Add globals & fns to define the board's design
All oxo boards are squares, and the board state is an array of:
empty  = - (hyphen)
cross  = X (upper case x)
nought = O (upper case o)

A lot of logic would get built around fit_board_to_square_grid, as we
will be able to play to the strengths of bash text processing + pattern
matching / REGEXP-fu. For example, a player wins a game if the pattern
of _any row_ is "X X X" or "O O O". This is simply a grep with return 0:

    grep -E "^X[[:space:]]+X[[:space:]]+X$"

To check if any a column is full; we can transpose the board, and
perform the grep against the transposed board.

To check if a diagonal is full; we can extract (say items 1,5,9 from the
game_state array) and once again grep against the same pattern.