Add fns for win/lose 'scoring' logic driven off a
regular, well-formatted, grid-based layout of the game state.

We base it on a regular, well-formatted, grid layout of the game state,
so we can have some fun exploiting the text processing power of bash.

We have three victory scenarios that we can test as follows.

Row Victory:

- A player wins a game if the pattern of _any row_ is "X X X " or
  "O O O " in our scheme of things. This is simply a strict grep that
  succeeds (exit code 0) against our grid view:

      grep -E "^X[[:space:]]X[[:space:]]X[[:space:]]$"

Column Victory:

- If we transpose the board (columns -> rows), then we can reuse the
  same grid-based grep-check to see if a player has won a column.

Diagonal Victory:

- And finally, if we extract diagonal items from the game state array
  into a grid-formatted row (say items 1,5,9), then we can again use
  the exact same grep test.