Prep for gameplay. Create row/column semantics and
apply them to creating a lookup table that we will use to translate
player input to a board position.

   e.g. a1 -> 1st position, b3 -> 6th position, c2 -> 8th position

Once we know a board position, we can set that position to the current
player's symbol X/O (if the position is empty).

As a bonus, the row/col semantics also help simplify and generalise
display formatting of the board, for player reference.